Begun work on full webserver
