+++ src/ds/web/Server.java
@@ -12,24 +12,10 @@ public class Server {
     public Server(int portnum) throws Exception {
         System.out.println("[Setup] Starting FlyAway Server.....");
         HttpServer server = HttpServer.create(new InetSocketAddress(portnum),0);
-        server.createContext("/", new MyHandler());
+        server.createContext("/", new RootHandler());
+        server.createContext("/query", new QueryHandler());
         server.setExecutor(null);
         server.start();
         System.out.println("[Setup] FlyAway Server started.");
-    }
-
-    static class MyHandler implements HttpHandler {
-        @Override
-        public void handle(HttpExchange t) throws IOException {
-
-            // Print Debug info for request
-            System.out.println("[Info] New Request, origin \"" + t.getRemoteAddress().getAddress() + ":" + t.getRemoteAddress().getPort() + "\", target \"" + t.getRequestURI() + "\"");
-
-            String response = "200 OK";
-            t.sendResponseHeaders(200,response.length());
-            OutputStream os = t.getResponseBody();
-            os.write(response.getBytes());
-            os.close();
-        }
     }
 }
